Move date-related methods from ST HomeController -> ContractStart step
I need access to the parsed contract_start_date and hire_date in order
to do some higher level validation. Moving these methods into the step
will achieve this without adding any duplication.

I've chosen to obtain all the inputs to the Calculator constructor from
the ContractStart step attributes for consistency.

I had to rescue ArgumentError exceptions in the step model versions of
the methods, because at this point we can't guarantee anything about the
values of the day, month & year attributes for each date.

I had to add the `looking_for` parameter to the controller spec to make
it more realistic, so that it was properly constructing the journey i.e.
it has the relevant value for `previous_step`.