Add specs & improve date-related methods on ST ContractStep model
Using Date.parse with an interpolated String gives more consistent
behaviour than Date.new and calling String#to_i on each field. In the
latter case, a date with a valid day and month, but no year, was
constructed as a date in the year 0000, rather than raising an
ArgumentError which is what one might've expected.