Rename ContractStart step & view -> TempToPermCalculator
The name of this step is left over from a time when it was only asking
the user for the contract start date and when it was in its own journey.
Now that it's asking the user many more questions and it's part of the
SupplyTeacher journey, it's useful to have a more specific name.