Add holidays to temp-to-perm calculator
Allow users to optionally enter dates of up to two school holidays that
appear between the contract start date and hire date. These will affect
the calculation of working days and therefore potentially affect the fee
that might be payable to the agency, as well the ideal hire and notice
dates, and the notice date based on the actual hire date.

This mirrors the behaviour in the Excel version of this calculator.