Error and maintenance page creation
We want to make our error and maintenance pages robust, one
way to achieve this is to minimise their dependencies on the
rest of the system.

This commit adds a rake task that creates
static pages that can  be served without the rest
of the system - all css, fonts and images are inlined. We use
the npm tool 'juice' to automate this.

To make it easier to keep these pages up-to-date with our
templates, we add an error_controller that can generate
these pages in a dynamic way, using layouts and the asset
pipeline.