Add anchor for generic hire_date validation error
This should've been added in this commit [1].

Note that there's no easy way to add the `govuk-input--error` CSS class
to *all three* `hire_date` fields, so I haven't done so for now. I
believe @chrisroos is working on adding more validation for the
combination of the three date fields and so I suspect he'll have to deal
with this problem then.

[1]: d16dcb86e214f6402b07326320f9e47da4821519