Validate presence of hire date
As per the guidance in the GOV.UK Design System[1]: "Check the entire
date, not the day, month and year separately."

The `hire_date` is a `Date` constructed from the day, month and year
parts. If it's present then we know it was a valid date. If it's absent
then it's either missing or an invalid date.

[1]: https://design-system.service.gov.uk/components/date-input/#error-messages