Disable/enable upload routes with environment variables.
We want to disable uploading routes for internet-facing
deployments of the application to provide better security.

This commit ensures that the routes for uploads are only
configured if APP_HAS_UPLOAD_PRIVILEGES is set.

I decided to make this check in config/routes because when
we talk about the change, we talk about it as enabling/disabling
routes rather than protecting controllers.

It required slightly more fiddly testing than I would have liked
because we have to ensure that the routes are reloaded before
and after controlling the ENVVAR.