Extract SupplyTeachers::Rate::TERMS into static record
This is more consistent with how we're doing this kind of thing
elsewhere in the codebase.

Things are slightly confusing, because there's already a
SupplyTeachers::Term static record, but that appears to be slightly
different - it's providing values for a UI select input which are then
mapped to the RateTerm values.