Add pro rata supplier day rate to temp-to-perm results page
In an attempt to clarify how the fee is calculated when the worker works
fewer than 5 days per week.

I wonder whether it'd be clearer to adjust the number of days
required/given (e.g. working 2 days a week would mean that 8 days notice
are required - 4 weeks x 2 days per week) instead of adjusting the day
rate. I've started exploring this in another branch but it's taking
quite a while so I'm making this change for now.