Use SupplyTeachers::JobType vs SupplyTeachers::Rate::JOB_TYPES
This removes some duplication, although it brings the overloading of
the meaning of SupplyTeachers::Rate#job_type into starker contrast.

I've had to add the non-role-related job types to the CSV data for
JobType and then add a new field to distinguish them from the
role-related job types. This isn't ideal, but I think it makes the code
more explicit and consistent across the piece and so is a small step
forward.