Correct logic in SupplyTeachers::Rate#term validation
A rate should have a term set if its job type is not one of the
role-related ones. Although previously the return value from
Rate#term_required? was the wrong way round, I'd achieved the correct
validation behaviour by also have the if/unless keys in the validation
definition the wrong way round!