Give dev & test users a Cognito sign-in option
As no one can get DfE accounts, they need to be able to use Cognito to
sign in to test the service.

This is implemented as a "secret" (not really secret, because the code
is open!) route:

    /supply-teachers/cognito

This adds a "Sign in with beta credentials" link to the gateway page
even when SUPPLY_TEACHERS_COGNITO_ENABLED is not set.