Use radio buttons for term in agency_payroll step
It seems inconsistent to use radio buttons for the job type (changed in
the previous commit) and a drop down for term, so I've decided to switch
this to use radio buttons too.

I think it's justified by the guidance in the GOV.UK Design System[1]:

> The select component should only be used as a last resort in
public-facing services because research shows that some users find
selects very difficult to use.

It's also easy enough to revert this commit if we decide that we do want
a drop down for this question.

[1]: https://design-system.service.gov.uk/components/select/#when-to-use-this-component