Make headings of fields in calculator <label>s
The accessibility audit picked us up on not having <label>s for
some of the fields in the calculator.

For those fields without <label>s, I've made the question text
the <label> for the field. I'm not entirely sure this is correct,
but it is the quickest thing we can do to be compliant without
any redesign. Is the real problem that these fields don't have
any labels?