Display a range in temp-to-perm fee results
If hiring between 9 and 12 weeks but not specifying a notice date then we
only know a range of possible fees.

Note that I've updated the feature specs so that they simulate a user
entering 4 weeks notice. This is to avoid having to update the
expectations to cater for the new fee ranges that can be displayed.