(fix) Add a fallback sign out route if there is no current session
Trello card: https://trello.com/c/nTwYhW9n/909-exception-when-signing-out-via-cognito

This PR fixes the "second sign out error" as detailed in the card. The issue was arising when a user logged out of Cognito, got an error (but was successfully signed out), hit the Back button in their browser, and clicked sign out again.

Because they had successfully signed out of Cognito, their session had been cleared. The auth controller was relying on a session to know where to redirect the user. This provides a fallback so that if there is no session pesent, and no current_login, a user hitting Sign Out will still be redirected to a valid page instead of seeing a 500 error.