(fix) Temp-to-perm calculator dates now require a year
We had an issue with using `Date.parse("YYYY-MM-DD")`, if you leave the
year off it defaults to this year. Our tests that check for the presence
of 'year' were passing incorrectly because they were using a single
digit month for the day, which combined with the absence of the year
produced an ArgumentError as expected. eg. `Date.parse("-1-1")` is invalid
as our tests expected but `Date.parse("-1-11")` is valid

To get the tests failing in the right way I had to change the day used
for each fixture to a double digits.

We then fixed the check for the presence of year by telling the `Date`
object which format we expect the date string to be in. Now if you omit
the year you get an ArgumentError as expected.