https://pushback.atlassian.net/browse/ST-11
ST-11 Results page content change

[DfE's request]
I’ve also noticed that the last paragraph of the results page for the TTP calculator still needs to be changed. We would like the statement to make it clear that by taking particular actions the school can pay 0.00 for taking the agency worker on. The current statement of:
Give notice to the agency on Thursday 28 February 2019 and make the worker permanent on Thursday 28 March 2019 so they've been in post for over 12 weeks and you provide 4 weeks’ notice.

Should be changed to:

If you give notice to the agency on Thursday 28 February 2019, and make the worker permanent on Thursday 28 March 2019 you will be charged £0.00 and avoid paying any temp­ to ­perm fee.

The £0.00 is important as it is immediately noticeable by the user and self­ explanatory.