MC & LS ITHC Remediation - User Account Enumeration changes
User Account Enumeration - The application was found to present error messages that facilitated the enumeration of valid user accounts.

Recommendation: BSI recommends that the application be configured to return generic error messages to ensure that they do not indicate whether a valid user account has been provided or not.

Current validation messages display the following: -

When password is wrong: Incorrect username or password.When user email is wrong: User does not exist.When resetting password and putting the wrong email in: Username/client id combination not found.

First two have been amended to always say 'Incorrect username or password'
The second one has been amended so it does not show an error message, but it now redirects to he password reset page where you see a message that if you have an account, you will receive instructions.