1: Make the migrations postcode_locator index operate only if it doesn't exist (and if it does when going down) 2: building-model - ensure an integer gia is put into the table row when take from the existing json data column 3: added the postcode_locator index to the postcode.rake file - which is called from these rake tasks     postcode     webpostcode     pctable 4: Added the command db:pctable to:     .travis.yml - in the script after db:test:prepare     pipeline_run_build_tests_sh, also after the db:test:prepare command     bin/docker-entrypoint.sh - after the bundle exec rake db:prepare command     bin/setup - after the 'system! 'bin/rails db:setup' command     bin/update - after the 'system! 'bin/rails db:setup' command 5: Moved the OrdnanceSurvey (postcode) creation and import code to a module 6: Included that module in postcode.rake and db:rake 7: Included the postcode index and view creation in the db:setup task
The intent is to have an index migration for the index that will do nothing (and be registered as a successful migration) if the index already exists and to have rake tasks that will create the index and views in its place.
The migration will remove the index on rollback however.

Then, as a rake task, db:pctable, to add the index and required postcode lookup views to the table
To call db:pctable everywhere it seems appropriate - specifically, when setting-up the db, setting up and updating the runtime environment, when preparing tests, as part of the travis build routine and during the instantiation of the environment within docker.

I think this covers all the bases where the opportunity to create these elements, which are kind-of outside of the application model-space, exists in the normal application runtime, on the desktop and within the Docker on AWS and during db:setup.