ensure that services ending with double digit codes can be found
The issue lay with the regex to identify if the spreadsheet was dealing
with a service or a subservice, eg MCF2.3.1 is a service but MCF2.3.1.1
is a subservice. Unfortunately the regex ending for checking ended with a d
rather than a d+ so it was only checking for single digits. The check
can now check for subservices ending with double digits as well