Add the supplier details section for RM6232
As the supplier details section for RM6232 is almost identical to the one in RM3830, I decided to make a new controller that both frameworks could use.
I also made them share as many as the validations on the model by creating a `SuppliersAdmin` concern.

I made sure to add specs for the controller using both frameworks and adding feature tests for RM6232 aswell to make sure nothing had broken.
I also added accessibility tests for the new pages.