I have added the ‘Start your procurement page’ for the new FM framework
While doing this I decided to update the step-by-step navigation as it has been updated. However, as this is not a component within GOV.UK Frontend we need to do the updates ourselves and on this front I have good news.

By looking through some documentation I was able to discover https://github.com/alphagov/govuk_publishing_components. This contains the CSS and Javascript we need to keep it up to date. It also includes Rails helpers to generate the content but I think it is too late in the day with this project to use it.

When giving the step-by-step navigation its fresh set of paint I updated all the other pages with step-by-step navigations.