Because this section is going to be a lot of work and require a lot of changes, I’ve decided to split the PRs up into more manageable sections.
For the contract details I have added a new controller/concern in order to deal with editing and showing the summary of the sections. This allows us to move a lot of the clutter out of the procurements controller and will allow code to be more easily shared.

I also added the specs for the work I have done so far which will need to be updated when further development is done