One of the biggest issues I have found with Facilities Management is how we work with buildings. There is a lot of duplicated code and things done in odd or non standard ways.
To make things easier for working in buildings on the new framework I have refactored the buildings controllers. We now have the key code defined by a single controller/concern which can be shared by the 3 current and 1 future controllers that concern editing buildings.

In the various buildings controllers, we now defined only the paths which are different for the different areas of the application.

As always, I have added tests to make sure these changes work. I have also updated the feature tests so that there is a buildings section for both frameworks.