As this is a rather complex section of the application, I’ve split it up into parts so that the PRs/commits do not get too large.
This commit concerns the adding of the ‘edit buildings’ section within the procurement for the new FM framework. I’ve done the minimum to get the ‘buildings’ page working in the details section and then added the ‘edit buildings’ section.

Included in this commit are tests for the ‘edit buildings’ section (both specs and feature tests) to make sure it works as expected.