This PR contains the changes required to get the edit ‘buildings’ page working.
Users are able to select and paginate the page to find all their buildings. This commit does not include the summary page which will be tackled in a future commit.

As part of this I have created a new table for the RM6232 procurement buildings. Although it is similar to the one in RM3830, it does not need to do as much so I have trimmed it down.

I have also updated the specs to take account of the changes made.