Add the ability to add services to a building in a procurement. This includes the validation of the service selection which should match the validation in the previous framework.
I also added the method to check the service selection so that part is complete. In the next PR I will add the code to make sure all the statuses are complete before continuing from the further service requirements page.

As well as adding some more specs, I also added some feature tests and some accessibility tests for the ‘Assigning services to buildings’ section. I also moved some of the step definitions about to make sure any that are shared between frameworks are in a neutral file.