Now that the details section for the new framework is complete, this commit deals with some features that don’t really fit into any particular area.
The first change was to make sure that if the user changed the service codes in the procurement, this would be updated in the procurement buildings (specifically when they remove an already selected service code).

The second change was to make sure that the user can add the missing region for a procurement building if an active procurement building has a missing region. As part of this I made most of the procurement_buildings controller shared for both frameworks which stop some repetition of code.

Finally, I add specs for all the changes and updated all the feature tests to make sure they worked.