In this commit I have added the results page for the new framework. As part of this I have also updated the specs and added feature tests.
One feature of transitioning to results is the requirement to freeze data that can be changed. For now I have just foxen the building data. We will need/want to freeze the supplier data so that we can have a record of it when the user transitioned to results. However, as this data model is not yet set, I will implement this change at a later date.