We are making major changes to the journey by removing everything after ‘What happens next?’ page. In this commit I have updated the ‘What happens next?’ page to the ‘What do I do next?’ page.
I have then removed the routes we are no longer using. Because there is a possibility that the work I’ve done so far may used in the future, I’ve not deleted anything but either moved it or commented it out. This should make reverting back to the current version fairly simple if we ever need to.