We are making major changes to the journey by removing everything after ‘What happens next?’ page. In this commit I have updated the ‘Your account’ page.
As we no longer need the user to enter buildings, I have removed the routes for the buildings in the RM6232 section.

Because there is a possibility that the work I’ve done so far may used in the future, I’ve not deleted anything but either moved it or commented it out. This should make reverting back to the current version fairly simple if we ever need to.