There is a requirement from the admin team that all changes that are made to the supplier data are logged. In this commit I have added the models which will store these changes along with who made them.
The first model is the `Admin::SupplierData`. This has a JSON of the full supplier data before it is imported into the database after a user uploads the data. We then have the `Admin::SupplierData::Edit` model which contains all the changes made since the most recent `Admin::SupplierData` was created.

As always I have added specs to make sure everything is working as expected. In a future commit I will add a section to view who made the changes.