There is a requirement from the admin team that all changes that are made to the supplier data are logged.
In this commit I have added the ability to download the supplier data at any instant in time. This may be needed for auditing purposes if, for example, a supplier complained that they were not being included in any searches.

As always I have added specs to make sure everything is working as expected.