There is a requirement from the admin team that they are able to download a management report so they can see who is using the service.
In this commit, I have updated the admin frontend so that they are able to view all generated reports and generate new ones. In a future PR I will add the ability to generate the report for the new RM6232 framework.

As always I have added specs to make sure everything is working as expected.