I have decided to convert all the JavaScript in the application to TypeScript.
We have quite a lot of JavaScript in this application and I do find it quite difficult to maintain. I am hopeful TypeScript will make all the maintenance easier and help future proof our code. It also gave me an opportunity to review our code and remove anything that was no longer needed.

(I also had nothing else to do at this moment in time and wanted to give myself a challenge)