I have refactored our GitHub actions to allow us to trigger deployments.
I have removed the existing scripts related to the build_test on AWS. I have also moved the unit-test and feature-test jobs into separate workflows so they can be reused in other workflows.

I have created the workflow which should (hopefully) then deploy the code to the correct environments, but this still needs to be tested and sorted.