We are adding the ability for the CSC team to manage users through the application rather than in cognito. This commit contains the code that, once the frontend has been developed, will allow them to do so.
I have created the `Cognito::Admin::User` model to handle the users we find in a more rails way. I have created the `Cognito::Admin::UserClientInterface` as an interface to handle the Cognito API, and keep some of the complexity out of the `Cognito::Admin::User` model. This interface allows us to:
- Create users
- Search users
- Find a single user
- Update the users attributes

In addition to these, I have created the  `Cognito::Admin::Roles` class which handles the logic for the roles that can be assigned to users by users. This is because we need to limit the ability of who can edit who based on their permissions. For example, most CSC users will only need to manage buyers so they do not need to be able to create Service admins.

As usual, I have added tests to make sure the changes I have implemented work.