We want to use Power BI to better analyse the FM data. To do this we need to regularly export the FM data and upload it to S3.
I have added a new SideKiq worker which can get the data export and do this upload. I’ve also added the sidekiq cron job which we can use to schedule the job every week.

I have also refactored the `ManagementReportCsvGenerator` in the new framework to use the `ProcurementCsvExport` like in the old framework. This way we can generate the report without needing to have a management report.