As the RM3830 Facilities Management framework needs to be closed soon, we need to add that ability to the code base. Fortunately, this is something that already exists in the Legacy code base so I was able to basically copy and use that.
Instead of having a `FacilitiesManagement::Framework` model, we instead have the `Framework` model with the `facilities_management` scope. I’ve updated the references where this was previously used to reflect these changes.

Finally, I’ve updated some of the authentication controllers so that they can all make use of the framework status and do the right actions if there is an issue with the frameworks.